move basic auth logic to auth file
